+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * added CRAN DOI badge
 * added 2021 typhoon data
+* added utility unexported function to download cyclone reports
 
 
 # bagyo 0.1.1